refactor: rename public folder to front to avoid PHP reserved keyword
- Renamed public/ folder to front/ to avoid PHP reserved keyword conflict
- Updated composer.json autoloader mapping from PdcConnector\Public\ to PdcConnector\Front\
- Renamed PublicCore.php to FrontCore.php
- Regenerated autoloader with composer dump-autoload
- Maintains clear separation between admin and user-facing functionality
