+++ bin/stop-wordpress
