+++ test/e2e/.auth/user.json
@@ -1,8 +1,8 @@
 {
   "cookies": [
     {
-      "name": "wordpress_d1c9e11eb00bdf205a8c6f83abac0894",
-      "value": "admin%7C1753681965%7CXVlkifdVNyHKDvCORViatfdwRDgZYrANGhs7LxfGPkh%7C8b34b3eda47dea79aa4b0f68edb3411067dc04775829136503dde6c213350395",
+      "name": "wordpress_db5d1b4ee4780d1f195600eee7ffc3a9",
+      "value": "admin%7C1753873858%7C70zTxQOLidDH77C6XQjOPWScfYUgtHznFvKVIfzCyWG%7C8fdd82c95481def60217455d71a185a415f500ed3aea013e9dc5c2ab25c55db4",
       "domain": "localhost",
       "path": "/wp-content/plugins",
       "expires": -1,
@@ -11,8 +11,8 @@
       "sameSite": "Lax"
     },
     {
-      "name": "wordpress_d1c9e11eb00bdf205a8c6f83abac0894",
-      "value": "admin%7C1753681965%7CXVlkifdVNyHKDvCORViatfdwRDgZYrANGhs7LxfGPkh%7C8b34b3eda47dea79aa4b0f68edb3411067dc04775829136503dde6c213350395",
+      "name": "wordpress_db5d1b4ee4780d1f195600eee7ffc3a9",
+      "value": "admin%7C1753873858%7C70zTxQOLidDH77C6XQjOPWScfYUgtHznFvKVIfzCyWG%7C8fdd82c95481def60217455d71a185a415f500ed3aea013e9dc5c2ab25c55db4",
       "domain": "localhost",
       "path": "/wp-admin",
       "expires": -1,
@@ -31,8 +31,8 @@
       "sameSite": "Lax"
     },
     {
-      "name": "wordpress_logged_in_d1c9e11eb00bdf205a8c6f83abac0894",
-      "value": "admin%7C1753681965%7CXVlkifdVNyHKDvCORViatfdwRDgZYrANGhs7LxfGPkh%7Cf0a4bcda0a3cb92b2dccd06313035f69beed0925df4996fd1b9d0d11d5f4d5d0",
+      "name": "wordpress_logged_in_db5d1b4ee4780d1f195600eee7ffc3a9",
+      "value": "admin%7C1753873858%7C70zTxQOLidDH77C6XQjOPWScfYUgtHznFvKVIfzCyWG%7Cf7015815346041bc0831813e349a2dcf7bb7452296ed99e980f334a3eb4872e6",
       "domain": "localhost",
       "path": "/",
       "expires": -1,
@@ -41,21 +41,31 @@
       "sameSite": "Lax"
     },
     {
+      "name": "tk_ai",
+      "value": "woo%3A8Ut41%2F47tFidWfb6sunmXg6p",
+      "domain": "localhost",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
       "name": "wp-settings-1",
       "value": "mfold%3Do%26libraryContent%3Dbrowse",
       "domain": "localhost",
       "path": "/",
-      "expires": 1785045165.690432,
+      "expires": 1785237058.38545,
       "httpOnly": false,
       "secure": false,
       "sameSite": "Lax"
     },
     {
       "name": "wp-settings-time-1",
-      "value": "1753509165",
+      "value": "1753701058",
       "domain": "localhost",
       "path": "/",
-      "expires": 1785045165.690473,
+      "expires": 1785237058.385491,
       "httpOnly": false,
       "secure": false,
       "sameSite": "Lax"
